[v1.0] remove deparse.c
The shell code de-parser, which converts byte code back to shell
source code, is unused. It was used in SHOPT_COSHELL, which was
removed in 3613da42, and in ancient code for systems without
fork(2), which was removed in 7b0e0776. Testing reveals it to be
quite broken as it has not kept up with more recent changes in ksh.

It is kept in the dev branch, as we intend to fix it up and use it
for 'typeset -f FUNCTIONNAME' to output function definitions in a
future release. (The current design, which outputs original source
code direct from the source file, is fundamentally broken because
the output of a function definition that was loaded from a file is
corrupted if you edit the file after loading the function.)