do not sh_close() a -1 file descriptor (re: 80767b1f, 411481eb)
Researching #483 revealed several instances in coprocess cleanup
where sh_close() is called with a file descriptor of -1 (which
flags that the pipe is already closed). As of feeb62d1, this sets
errno to EBADF. While the race condition triggered by this was
fixed in 411481eb, it's still better to avoid it.

This re-applies most of the changes reverted in 80767b1f.