Fix race condition in coprocess test with external 'cat'
The race is between '$cat |&' and 'kill $pid'. In between, there
are only a variable assignment and two buffered writes, so there
is nothing that waits for the external 'cat' to finish forking,
execve'ing and initialising -- meaning there is no guarantee it is
ready to catch SIGTERM. This explains the hang; 'cat' misses the
signal, continues to initialise, and simply waits for more input.

src/cmd/ksh93/tests/coprocess.sh:
- Actually read from the /bin/cat coprocess and verify that it
  works. This has the beneficial side effect of ensuring it is
  fully loaded and initialised before SIGTERMing it.

Resolves: https://github.com/ksh93/ksh/issues/132