tests/leaks.sh: quadruple maximum number of test iterations
There were still intermittent leak test failures on Debian.
They're clearly spurious as a different test fails each time.

My testing on a Debian 11.4.0 VM showed that it takes not a
doubling, but a quadrupling of the maximum number of test
iterations (from 16384==2**14 to 65536==2**16) for that system's
memory state to stabilise enough to avoid spurious test failurese.

Resolves: https://github.com/ksh93/ksh/issues/501