Revert "[v1.0] Remove experimental .getn discipline"
This reverts commit 42fc5c4c0da82b4427c34f2e87a76e0171e560b4.
It somehow caused the following reproducer to fail. Reason unknown.

  typeset -i x
  function x.set { :; }
  x[0]=0
  unset x
  typeset -p x

Expected output: none

Actual output:

  typeset -a -i x=()

The 'x' variable fails to be unset.

With the .get and .getn discipline functions instead of .set, the
above reproducer still fails even after the revert, but that always
has failed, back to at least 93u+ 2012-08-01.

This is yet more evidence that discipline functions are a mess.
But a proper cleanup will require a lot of time and very thorough
testing, so it will have to wait until some future release.

src/cmd/ksh93/tests/variables.sh:
- In addition to the revert, add the above reproducer as a
  regression test for the .set, .unset and .append disciplines.