vi mode: Disable = and tab at the start of the line (#521)
In vi mode, due to a buffer overflow, <ESC>= as the first input
either produces an invalid completion or crashes the shell. The
easiest fix is to disable <ESC>= as well as <TAB> completion at
the start of the command line, as is already done in emacs mode.

src/cmd/ksh93/edit/vi.c:
- getline(): If cur_virt <= 0 in case '\t', beep and refuse tab
  completion.
- textmod(): Move an 'if' statement checking for INVALID cur_virt
  to include = mode along with * and \ mode.

Co-authored-by: Martijn Dekker <martijn@inlv.org>
Resolves: https://github.com/ksh93/ksh/issues/520