No completion before 1st char or in empty line (re: b6c8bb7b) (#523)
There are still two contexts in which the unhelpful "alias +
builtin + path" completion occurs:

* After the start of an empty line
* After the start of a line, but before the first character

To stop completion from being attempted in these contexts, a loop
is added to vi.c and emacs.c that looks over the previous positions
(current position inclusive) and allows the completion process to
continue once it finds the first non-space character. The existing
placement restrictions remain in case the cursor arrives in a
negative position via a bug.

Co-authored-by: Martijn Dekker <martijn@inlv.org>