man page: fix confusing use of "null" for empty string
"Empty string" is the term typically used in the POSIX shell
universe for string values of zero length. But the manual page uses
"null string" or "null value", which can easily be confused with
the variable being null as in unset, having no value -- a state
that is distinct from it having the empty string value.

Thanks to @jghub for pointing this out.