nvtype.c: fix erroneous alignment calculation
_Align_ is a dummy struct that is used to calculate alignment
offsets in nv_datasize() via the alignof macro. The _cint32_t and
_dint32_t struct members, added in ksh 93u+ 2012-05-15, do not
match the pattern of the pre-existing entries: the _c* members are
all of type char. And none of the _d* members except pointer are
pointers, but the new _dint32_t member is a pointer. Which does not
match the size of int32_t which is actually what's being queried on
the line s = sizeof(int32_t);.

This has been identified as the cause of a seemingly random crash
in a script using 'typeset -T' types in combination with short int
or enum type members.

src/cmd/ksh93/sh/nvtype.c:
- _Align_: Make the new _cint32_t a char and remove the pointer
  asterisk from _dint32_t, making them fit the other entries and
  allowing correct alignment calculation.
- Fix another inconsistency. Short integers (typeset -si) are typed
  in C as int16_t, not short. On most (all?) current systems those
  are the same size, but it's not guaranteed. Fixed by adding
  corresponding _cint16_t and _dint16_t memmbers to _Align_ and
  using {alignof,sizeof}(int16_t) in nv_datasize().
- Remove now-unused _{c,d}{short,long} _Align_ struct members.

Thanks to Janis Papanagnou for finding and reporting the bug.
Resolves: https://github.com/ksh93/ksh/issues/537