Fix LINENO being wrong after being set in an invocation-local scope
LINENO is reset after being used in an invocation-local scope
(temporary assignment preceding a command). Reproducer:

	:
	LINENO=$LINENO true
	print "Line 3 is $LINENO"

Output: Line 3 is 1
Expect: Line 3 is 3

The bug was introoduced in ksh 93u- 2010-10-26 by adding LINENO to
the variables listed in nv_cover() in init.c. That function is used
to determine if a variable's discipline function (nvfun) should be
copied over when making a locally scoped version of it. The 93u-
changelog does not mention why this change was made.

Reverting it works for 93u+, but causes the above reproducer to
crash in 93u+m. A related 93u+m crash occurs with $RANDOM:

	$ ksh -c 'RANDOM=$RANDOM true; test -n $RANDOM'
	Memory fault

These crashes were introduced in 75b247cc by this line in name.c:
566:		mp->nvfun = np->nvfun;

Copying the variable's discipline function pointer nvfun should
only be done if nv_cover() indicates that it should be done.

src/cmd/ksh93/sh/init.c: nv_cover():
- Revert the 2010-10-26 addition of LINENO.

src/cmd/ksh93/sh/name.c: nv_setlist():
- For invocation-local scopes, only copy nvfun if nv_cover()
  indicates it for that variable, otherwise reset it.

Thanks to @JohnoKing for reporting and researching the bug.
Resolves: https://github.com/ksh93/ksh/issues/543