libast: printf: add missing mbwidth() result checks
In ksh93u- 2010-08-11, the mbwidth() macro was changed so that it
returns -1 for a control character (which has no width) or an
invalid multibyte character. But the uses of mbwidth() in
sfvprintf.c were not updated to check for this. As a result, byte
offsets were corrupted, causing something like 'printf %Lb "\0200"'
to intermittently output garbage.

src/lib/libast/sfio/sfvprintf.c:
- Add missing checks for negative mbwidth() result.
- Gets rid of all the #ifdef mbwidth directives, as we are working
  with a known version of libast.

This fix extends and somewhat refactors a patch from Apple:
https://opensource.apple.com/source/ksh/ksh-25/patches/src__lib__libast__sfio__sfvprintf.c.diff.auto.html

Related: https://github.com/att/ast/issues/14
Resolves: https://github.com/ksh93/ksh/issues/544