Fix build with link time optimisation on macOS
When building with -flto in CCFLAGS and/or LDFLAGS, the following
linking error occurs upon linking ksh with libshell.a:

  ld: reference to bitcode symbol '_setenv' which LTO has not
  compiled in '_procopen' from /tmp/lto.o for architecture x86_64

I'm not sure exactly how, but this seems to be caused by a hack to
forcibly override the system's setenv(3) in comp/setenv.c to use
AST setenviron(3). This hack is for the benefit of one single
setenv call in misc/procopen.c.

src/lib/libast/comp/setenv.c:
- Do not forcibly override the system's setenv regardless of the
  iffe test result for its presence.

src/lib/libast/misc/procopen.c:
- Provide a copy of the setenv AST fallback version from
  comp/setenv.c as a local/static alternative function under a
  different name.
- Call that instead of setenv(3).

Thanks to Jeffrey H. Johnson (@johnsonjh) for the report.
Resolves: https://github.com/ksh93/ksh/issues/552