Fix BUG_BRACQUOT remnant (re: e80c2650, 3eff4f54)
There is a missing check for single or double quotes when parsing
the terminating "]" character in a bracket expression. Thus, a
quoted "]" resets the flag that enables the translation of quotes
into backslash escapes, reintroducing the bug in that context.

Example reproducer:

  case b in ["]"a"-"z]) echo BUG;; esac

src/cmd/ksh93/sh/macro.c: copyto(): case S_ENDCH:
- If the current character cp[-1] is a ']', then before decreasing
  bracketexpr, also check that it is not in a single or double
  quoted context.