INIT: remove unused 'execrate' utility
Apparently, Cygwin around 2002 had a problem where it was not
possible to invoke commands without their .exe file name suffix,
e.g. you had to say /bin/mv.exe instead of /bin/mv. The execrate
wrapper script was written to work around the problem for specific
commands used by the build system. It includes a sanity test that
is executed when 'execrate' is invoked with no arguments, and that
exits with status 0 if the workaround is needed; the utility is
only used if the test exits 0. It exits 1/false on every system I
can get my hands on, including a decade-old version of Cygwin on
Windows 2000. So I think we can safely call this obsolete.