init: radix point and other tweaks
The way the radix point ('.' or ',') is initialised is potentially
problematic as we're importing environment variables via env_init()
before it is set. If the environment never sets a locale, then
SHLVL, which has a default NV_INTEGER attribute, is converted to an
integer which involves parsing an arithmetic expression, which may
fail with a spurious syntax error if sh.radixpoint is not yet set.
(It is also potentially a problem when importing type attributes
from the environment, which we're still doing on the 1.0 branch.)
The fix is to pre-set sh.radixpoint to '.' before calling env_init.

src/cmd/ksh93/sh/init.c:
- init_radixpoint():
  - Since this now won't be called from two places (see below),
    remove as a separate function and integrate in put_lang().
  - Remove checks for non-null pointer or empty string and trust
    that localeconv() returns a valid string pointer; POSIX says
    there are no error conditions.
  - Handle the hypothetical (?) case of a locale providing a multi-
    byte radix character by defaulting to '.' instead, as the arith
    code does not support it.
- sh_init():
  - Remove obsolete static 'beenhere' variable; it was left over
    from removed preparations to make it possible to have more than
    one shell state in the same process (re: 2d3ec8b6).
  - Replace init_radixpoint() call here with a preliminary
    init of the sh.radixpoint to '.' as described above.
  - Move set/unset of SH_INIT shell state to start/end of init.